requirements.txt
google-genai>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.29.0

